[A] topic validators, topic endpoints

+++ backendApp/src/main/java/com/internet_forum/springboot/controller/TopicController.java
@@ -6,6 +6,7 @@ import com.internet_forum.springboot.model.Topic;
 import com.internet_forum.springboot.model.UserEntity;
 import com.internet_forum.springboot.repository.TopicRepository;
 import com.internet_forum.springboot.service.TopicService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,7 +14,9 @@ import org.springframework.security.core.Authentication;
 
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
@@ -30,25 +33,40 @@ public class TopicController {
         this.topicService = topicService;
     }
 
+
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public TopicResponseDto createTopic(@RequestBody TopicRequestDto topic, Authentication authentication) {
+    public TopicResponseDto createTopic(@Valid @RequestBody TopicRequestDto topic, Authentication authentication) {
         UserEntity userEntity = (UserEntity) authentication.getPrincipal();
         Long userId = userEntity.getId();
         return topicService.createTopic(topic, userId);
     }
-
-    @DeleteMapping
-    public void deleteTopic(@RequestParam Long topicId){
-        topicRepository.deleteById(topicId);
-
-    }
     @GetMapping
-    public List<TopicResponseDto> getTopics(){
+    public List<TopicResponseDto> getTopics() {
         return topicService.getTopics();
+
     }
     @GetMapping("/{id}")
-    public Optional<TopicResponseDto> getTopicById(@PathVariable Long id) {
-        return topicService.getTopicById(id);
+    public ResponseEntity<TopicResponseDto> getTopicById(@PathVariable Long id) {
+        Optional<TopicResponseDto> topic = topicService.getTopicById(id);
+        return topic.map(response -> new ResponseEntity<>(response, HttpStatus.OK))
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    }
+
+
+    @DeleteMapping
+    public void deleteTopic() {
+        topicService.deleteTopic();
+    }
+
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteTopicById(@PathVariable("id") long topicId) {
+        if (topicRepository.existsById(topicId)) {
+            topicRepository.deleteById(topicId);
+            return new ResponseEntity<>("Topic deleted successfully", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Topic not found",HttpStatus.NOT_FOUND);
+        }
     }
 }
